requirements.txt
streamlit==0.88.0
matplotlib==3.4.2
pip==21.3.1
pandas==1.3.5
Pillow==8.2.0